requirements.txt
# Core web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Templates + static files
jinja2==3.1.4
python-multipart==0.0.9

# Environment variables
python-dotenv==1.0.1

# PyTorch (CPU by default, install torch nightly if on MPS)
torch>=2.0.0

# Hugging Face ecosystem
transformers==4.44.0
sentence-transformers==3.0.1
accelerate==0.33.0

# Vector DB
faiss-cpu==1.8.0

# PDF / OCR (if you’re ingesting docs)
pymupdf==1.24.2
pytesseract==0.3.10
pdfplumber==0.11.0

# Document processing
python-docx==1.1.2
unstructured==0.15.0

# Data wrangling
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
